amend wrong pydoc. move unet out from pix2pix generator modules.
