+++ networks/nets/unet/__init__.py
@@ -1,0 +1,2 @@
+from .u128 import UNet128Genarator as UNet128
+from .u256 import UNet256Genarator as UNet256
